+++ khoros/objects/__init__.py
@@ -6,10 +6,10 @@
 :Example:           ``objects.archives.archive(khoros_obj, '123', suggested_url, return_status=True)``
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     20 Dec 2020
+:Modified Date:     25 Mar 2021
 """
 
-__all__ = ['albums', 'archives', 'messages', 'roles', 'settings', 'users']
+__all__ = ['albums', 'archives', 'messages', 'roles', 'settings', 'subscriptions', 'users']
 
 # Import all submodules by default
 from . import albums
@@ -17,4 +17,5 @@ from . import archives
 from . import messages
 from . import roles
 from . import settings
+from . import subscriptions
 from . import users